refactor(capabilities): migrate the text cap to the struct-hosted actor macro
Move the `impl NativeActor for TextCapability` block into `text/runtime.rs`
under `#[runtime]` and lift `#[actor(singleton)]` onto the struct (#2382).
